SaveLoad versioning.
Metaprogram keeps track of all versions of a given type in SaveLoadVersions.txt. Changes in order, type,
type version, and constant value (only for enums) of fields cause version of enclosing type to be bumped.
Generated SaveLoad functions always save the latest version and branch on version for loading.
Unknown fields and types are loaded and immediately discarded. Once old versions are guaranteed
to not be needed, they can be manually deleted from SaveLoadVersions.txt to reduce the size of
the generated code.