Save/Load shader cache.
Shader bytecode and some metadata are stored on shutdown. On load we can read
and validate shaders against metadata of the current build and shader sources.

During loading all of the data is read into the temp allocator and only copied
to the heap when necessary. In release builds we could remove validation and keep
everything in the temp allocator to reduce the amount of copies.